faq-capture-actions/requirements.txt
rasa-sdk==2.3.1
requests==2.25.0
slack-sdk==3.4.2
python-dotenv==0.15.0